Explicitly return false for networks of different IP versions
It doesn't matter if there is a way to map IPv4 addresses into IPv6,
`::ffff:10.0.0.1` is never the same thing as `10.0.0.1`. It is
frustrating to me that the golang net package parses IP addresses in
IPv4 format into IPv6 mapped addresses. I want to be sure that these
utilities never accidentally compare them equally.