Conform to go guidelines for receiver names
The go wiki has this to say about a method's receiver name:

The name of a method's receiver should be a reflection of its
identity; often a one or two letter abbreviation of its type suffices
(such as "c" or "cl" for "Client"). Don't use generic names such as
"me", "this" or "self", identifiers typical of object-oriented
languages that place more emphasis on methods as opposed to functions.

https://github.com/golang/go/wiki/CodeReviewComments#receiver-names