refactor canCombineNets
First, I noticed that the initial comparison of IP addresses could be
a bug because the network portion of the IP could be the same but the
host part could be different. That could cause the method to determine
that the following two "networks" could be combined into 10.0.0.0/23,
which isn't true because they are the same network.

    10.0.0.0/24
    10.0.0.1/24

Also used NetworkAddr to ensure that the result does not have any ones
in the host part for good hygiene.

Then, I reorganized the method a little bit for readability.